+++ SDK/UI/View/HTBBookmarkRootView.m
@@ -198,7 +198,7 @@
 - (void)updateTextCount
 {
     NSInteger textCount = ceilf((float)[self.commentTextView.text lengthOfBytesUsingEncoding:NSUTF8StringEncoding] / 3.f);
-    self.textCountLabel.text = [NSString stringWithFormat:@"%d", textCount];
+    self.textCountLabel.text = [NSString stringWithFormat:@"%ld", (long)textCount];
     if (textCount > 100) {
         self.textCountLabel.textColor = [UIColor redColor];
     }
